server/requirements.txt
fastapi
uvicorn[standard]
pydantic
email-validator
python-dotenv
python-jose[cryptography]
boto3
